fix(types): change duration_ms type from Option<u64> to Option<f64> for improved precision

+++ crates/scrobbler/src/listenbrainz/types.rs
@@ -8,7 +8,7 @@ pub struct AdditionalInfo {
     pub release_name: Option<String>,
     pub musicbrainz_artist_id: Option<String>,
     pub musicbrainz_track_id: Option<String>,
-    pub duration_ms: Option<u64>,
+    pub duration_ms: Option<f64>,
     pub media_player: Option<String>,
     pub submission_client: Option<String>,
     #[serde(flatten)]
